@@ -1,13 +1,8 @@
 import { Component, Input, OnInit, OnDestroy } from '@angular/core';
-import { Subscription }   from 'rxjs/Subscription';
-import { Observable }        from 'rxjs/Observable';
-import { Subject }           from 'rxjs/Subject';
-import { BrowserModule } from '@angular/platform-browser';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { Subscription, Subject } from 'rxjs';
 import {trigger, state, style, animate, transition} from '@angular/animations';
-import 'rxjs/add/observable/of';
 
-import { NavigationServiceService } from '../navigation-service.service';
+import { NavigationService } from '../navigation.service';
 import {Kategorie} from '../model/Kategorie';
 import {Themenfeld} from '../model/Themenfeld';
 
@@ -17,7 +12,7 @@ import {Themenfeld} from '../model/Themenfeld';
   templateUrl: './themenfelder.component.html',
   styleUrls: ['./themenfelder.component.css'],
   animations: [
-    trigger('heroState', [
+    trigger('themenfeldTrigger', [
       state('inactive', style({
         backgroundColor: '#eee',
         width: '0px',
@@ -25,51 +20,49 @@ import {Themenfeld} from '../model/Themenfeld';
       state('active',   style({
         backgroundColor: '#cfd8dc',
          width: '300px',
-      })),      
+      })),
     transition('inactive => active', animate('1000ms ease-in')),
     transition('active => inactive', animate('1000ms ease-out'))
   ])
   ]
 })
 export class ThemenfelderComponent implements OnInit, OnDestroy {
-  
   private subscriptionKategorie: Subscription;
 
   private subscriptionThemenfeld: Subscription;
 
-  state: string = 'inactive';
+  state = 'inactive';
 
   themenfelder: Subject<Themenfeld[]> = new Subject<Themenfeld[]>();
-  
-  
-  
+
   ngOnDestroy(): void {
-    this.subscriptionKategorie.unsubscribe();
-    this.subscriptionThemenfeld.unsubscribe();
+    if (this.subscriptionKategorie) {
+      this.subscriptionKategorie.unsubscribe();
+    }
+    if (this.subscriptionThemenfeld) {
+      this.subscriptionThemenfeld.unsubscribe();
+    }
   }
 
-  constructor(private navigationService: NavigationServiceService) {
+  constructor(navigationService: NavigationService) {
 
     this.subscriptionKategorie = navigationService.selectedKategoryObs
-      .subscribe((kategorie : Kategorie) => this.kategorieChanged(kategorie));
+      .subscribe((kategorie: Kategorie) => this.kategorieChanged(kategorie));
 
     this.subscriptionThemenfeld = navigationService.selectedThemenfeldBs
-      .subscribe((themenfeld : Themenfeld) => this.themenfeldChanged(themenfeld));
+      .subscribe((themenfeld: Themenfeld) => this.themenfeldChanged(themenfeld));
   }
 
-  private kategorieChanged(kategorie: Kategorie)
-  {
+  private kategorieChanged(kategorie: Kategorie) {
     this.state = (this.state === 'inactive' ? 'active' : 'inactive');
     this.themenfelder.next(kategorie.themenfelder);
   }
 
-  private themenfeldChanged(themenfeld: Themenfeld)
-  {
-    this.state = "inactive";
+  private themenfeldChanged(themenfeld: Themenfeld) {
+    this.state = 'inactive';
   }
 
   ngOnInit() {
-    
   }
 
 }
