@@ -2,7 +2,7 @@ import { ContentHostDirective } from './content-host.directive';
 
 describe('ContentHostDirective', () => {
   it('should create an instance', () => {
-    const directive = new ContentHostDirective();
+    const directive = new ContentHostDirective(null);
     expect(directive).toBeTruthy();
   });
 });
