@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {NavigatorDataService} from '../navigator-data.service'
+import {NavigatorDataService} from '../navigator-data.service';
 import {Kategorie} from '../model/Kategorie';
 import {Navigator} from '../model/Navigator';
 
@@ -10,15 +10,14 @@ import {Navigator} from '../model/Navigator';
 })
 export class KategorienComponent implements OnInit {
 
-  public kategorien : Kategorie[];
+  public kategorien: Kategorie[];
 
-  constructor(private navigatorDataService : NavigatorDataService) {
-   }
+  constructor(private navigatorDataService: NavigatorDataService) {
+  }
 
   ngOnInit() {
-    this.navigatorDataService.getIndex().then((navigator : Navigator) => 
-      this.kategorien = navigator.kategorien);
-      
+    this.navigatorDataService.getIndex().then((navigator: Navigator) =>
+    this.kategorien = navigator.kategorien);
   }
 
 }
