@@ -1,6 +1,5 @@
 import {Box} from './Box';
 
-export class Boxes
-{
-    public boxes :  Box[];
+export class Boxes {
+    public boxes:  Box[];
 }
