@@ -1,11 +1,8 @@
-import {Themenfeld} from './Themenfeld'
+import {Themenfeld} from './Themenfeld';
 
 export class Kategorie {
 
     public name: string;
-    
     public isnew: boolean;
-
-    public themenfelder : Themenfeld[];
-
+    public themenfelder: Themenfeld[];
 }
