@@ -2,6 +2,6 @@ import {NavigatorContainer} from './NavigatorContainer';
 
 export class Page {
 
-    public pageId : number;
+    public pageId: number;
     public containers: NavigatorContainer[];
 }
