@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Headers, Http } from '@angular/http';
-import 'rxjs/add/operator/toPromise';
+import {HttpClient} from '@angular/common/http';
 
 import {Navigator} from './model/Navigator';
 import {Page} from './model/Page';
@@ -9,45 +8,27 @@ import {Page} from './model/Page';
 @Injectable()
 export class NavigatorDataService {
 
-  private headers = new Headers({'Content-Type': 'application/json'});
-  private url = './assets/navigation.json';  // URL to web api
+  private url = './api/navigation.json';  // URL to web api
 
-  private conenturl = './assets/page';
+  private content = './api/page';
 
 
-  constructor(private http: Http) { }
+  constructor(private http: HttpClient) { }
 
   public getIndex(): Promise<Navigator> {
-    return this.http.get(this.url)
+    return this.http.get<Navigator>(this.url)
                .toPromise()
-               .then(response => {
-                 let navigator = response.json() as Navigator;
-                 return navigator;
-                })
                .catch(this.handleError);
   }
 
-  public getConent(page: number) : Promise<string> {
-    let url = this.conenturl + page + ".html";
-    return this.http.get(url)
-        .toPromise()
-        .then(response => response.text())
-        .catch(this.handleError);
-  }
-
-  public getPage(page: number) : Promise<Page> {
-     let url = this.conenturl + page + ".json";
-      return this.http.get(url)
+  public getPage(page: number): Promise<Page> {
+      const url = `${this.content}${page}.json`;
+      return this.http.get<Page>(url)
                .toPromise()
-               .then(response => {
-                 let page = response.json() as Page;
-                 return page;
-                })
                .catch(this.handleError);
   }
 
   private handleError(error: any): Promise<any> {
-    debugger;
     console.error('An error occurred', error); // for demo purposes only
     return Promise.reject(error.message || error);
   }
