@@ -1,14 +1,28 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { of } from 'rxjs';
 
 import { PathComponent } from './path.component';
+import { NavigationService } from '../navigation.service';
+import { NavigatorDataService} from '../navigator-data.service';
+import { Kategorie } from '../model/Kategorie';
+import { Themenfeld } from '../model/Themenfeld';
 
 describe('PathComponent', () => {
   let component: PathComponent;
   let fixture: ComponentFixture<PathComponent>;
 
   beforeEach(async(() => {
+    const navigationServiceSpy = {
+      selectedKategoryObs: of(new Kategorie()),
+      selectedThemenfeldBs: of(new Themenfeld()),
+    };
+    const navigatorDataserviceSpy = jasmine.createSpyObj('NavigatorDataService', ['ba']);
     TestBed.configureTestingModule({
-      declarations: [ PathComponent ]
+      declarations: [ PathComponent ],
+      providers: [
+        {provide: NavigationService, useValue: navigationServiceSpy},
+        {provide: NavigatorDataService, useValue: navigatorDataserviceSpy},
+      ]
     })
     .compileComponents();
   }));
